# backend/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Apartment": {
				"description": "Apartment details",
				"properties": {
					"apartmentDescription": {
						"type": "string"
					},
					"roomCount": {
						"type": "integer",
						"format": "int32"
					},
					"surfaceArea": {
						"type": "number",
						"format": "double"
					},
					"streetAddress": {
						"type": "string"
					},
					"postalCode": {
						"type": "string"
					},
					"postDistrict": {
						"type": "string"
					}
				},
				"required": [
					"apartmentDescription",
					"roomCount",
					"surfaceArea",
					"streetAddress",
					"postalCode",
					"postDistrict"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"apartmentDescription": "2h+k+ph+parv.",
					"roomCount": 2,
					"surfaceArea": 54.5,
					"streetAddress": "Mallikatu 27 A 8",
					"postalCode": "33100",
					"postDistrict": "Tampere"
				}
			},
			"AuthToken": {
				"description": "AuthToken is used for granting access into dominus and identifying the user",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"username": {
						"type": "string"
					},
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username",
					"accessToken"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"id": 82,
					"username": "oskkos",
					"accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NCwidXNlcm5hbWUiOiJvc2trb3MiLCJpYXQiOjE1OTAyMjMxOTgsImV4cCI6MTU5MDMwOTU5OH0.tP-5h7NmH9aafoIKACrlkFZkLpY3slRRdJbTMk8jMHw"
				}
			},
			"AuthUser": {
				"description": "AuthUser is used for authenticating into dominus",
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"username": "oskkos",
					"password": "s3cret"
				}
			},
			"AddUserBody": {
				"description": "Payload for adding a new user",
				"properties": {
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"username",
					"password",
					"name"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"username": "oskkos",
					"password": "s3cret",
					"name": "Oskari Kosonen"
				}
			},
			"User": {
				"description": "Describes a user in a dominus",
				"properties": {
					"id": {
						"type": "integer",
						"format": "int32"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"username"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"id": 82,
					"username": "oskkos"
				}
			},
			"ChangePasswordBody": {
				"description": "Payload for changing password",
				"properties": {
					"oldPwd": {
						"type": "string"
					},
					"newPwd": {
						"type": "string"
					}
				},
				"required": [
					"oldPwd",
					"newPwd"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"oldPwd": "too_easy",
					"newPwd": "s3cret_p455word"
				}
			}
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"in": "header",
				"name": "x-access-token"
			}
		}
	},
	"info": {
		"title": "dominus-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/apartments": {
			"post": {
				"operationId": "AddApartment",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Apartment"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Add new apartment",
				"tags": [
					"Apartments"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Apartment: contains fields required for adding new apartment",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Apartment"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetApartments",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Apartment"
									},
									"type": "array"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Get all apartments of logged in user",
				"tags": [
					"Apartments"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": []
			}
		},
		"/apartments/{apartmentId}/addCoOwner/{coOwnerId}": {
			"put": {
				"operationId": "AddCoOwner",
				"responses": {
					"204": {
						"content": {
							"application/json": {}
						},
						"description": "No content"
					}
				},
				"description": "Add new apartment",
				"tags": [
					"Apartments"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [
					{
						"description": "Apartment that should get new co-owner",
						"example": 1,
						"in": "path",
						"name": "apartmentId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					},
					{
						"description": "New co-owner",
						"example": 2,
						"in": "path",
						"name": "coOwnerId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer"
						}
					}
				]
			}
		},
		"/auth/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthToken"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Authenticates user into dominus. Supply proper username and password.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "AuthUser: contains details required for signing into system",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthUser"
							}
						}
					}
				}
			}
		},
		"/auth/addUser": {
			"post": {
				"operationId": "AddUser",
				"responses": {
					"204": {
						"content": {
							"application/json": {}
						},
						"description": "No content"
					}
				},
				"description": "Add new user",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "AddUserBody: contains fields required for adding new user",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddUserBody"
							}
						}
					}
				}
			}
		},
		"/users/self": {
			"get": {
				"operationId": "GetSelf",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						},
						"description": "Ok"
					}
				},
				"description": "Entry point for getting self",
				"tags": [
					"Users"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": []
			}
		},
		"/users/changePassword": {
			"put": {
				"operationId": "ChangePassword",
				"responses": {
					"204": {
						"content": {
							"application/json": {}
						},
						"description": "No content"
					}
				},
				"description": "Entry point for changing password for current user",
				"tags": [
					"Users"
				],
				"security": [
					{
						"apiKey": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "ChangePasswordBody: contains oldPwd and newPwd",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ChangePasswordBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:7000/"
		}
	]
}